requirements.txt
openai>=1.70.0
pillow>=11.1.0